requirements.txt
Flask
slack_bolt
google-genai
google-auth
gspread
firebase-admin
llama-cpp-python
pytest
